Share OxyPlot.WindowsForms implemention with OxyPlot.Core.Drawing (#1529)
